Move linspace to .R
