+++ examples/mandelbrot.aya
@@ -1,18 +1,18 @@
 .# Also see examples/canvas/mandelbrot
 
-import ::matrix
-
 60 :width;
 27 :height;
 
-[-2 0.5 width.linspace] matrix! :x;
-[1 -1  height.linspace] matrix! .t :y;
-y :0i1 * x + :a;
+.# Create complex plane a
+[-2 0.5 width]  .R :x;
+[1  -1  height] .R :y;
+y :0i1 * x `+ :*   :a;
 
-aE matrix.zeros :z;
+.# Matching plane of zeros
+0 a:E L :z;
 
-.# Generate the fractal
+.# The mandelbrot algorithm
 z {2^a+} 20 % .|
 
 .# Display
-.rows {2<" #"\I} O "\n" % :P
+2 < " #" \ I "\n" % :P
